Rename class 'ManifestService' to 'EManifest'. This class will now be used to encapsulate intereactions with e-Manifest (the RCRAInfo module)
we're extracting the e-Manifest parts of the manifest service to help keep the separation of responsiblities clear
