+++ server/apps/trak/services/__init__.py
@@ -1,8 +1,5 @@
+from .emanifest import EManifest, EManifestError, PullManifestsResult, TaskResponse
 from .manifest_services import (
-    ManifestService,
-    ManifestServiceError,
-    PullManifestsResult,
-    TaskResponse,
     update_manifest,
 )
 from .waste_services import get_dot_hazard_classes, get_dot_id_numbers, get_dot_shipping_names
