+++ server/apps/trak/tasks/manifest_task.py
@@ -10,17 +10,17 @@ logger = logging.getLogger(__name__)
 @shared_task(name="pull manifest", bind=True, acks_late=True)
 def pull_manifest(self: Task, *, mtn: List[str], username: str) -> dict:
     """
-    This task initiates a call to the ManifestService to pull a manifest by MTN
+    This task initiates a call to the EManifest to pull a manifest by MTN
     """
 
     from apps.core.services import TaskService
-    from apps.trak.services import ManifestService
+    from apps.trak.services import EManifest
 
     logger.info(f"start task {self.name}, manifest {mtn}")
     task_status = TaskService(task_id=self.request.id, task_name=self.name, status="STARTED")
     try:
-        manifest_service = ManifestService(username=username)
-        results = manifest_service.pull_manifests(tracking_numbers=mtn)
+        emanifest = EManifest(username=username)
+        results = emanifest.pull(tracking_numbers=mtn)
         task_status.update_task_status(status="SUCCESS", results=results)
         return results
     except (ConnectionError, TimeoutError):
@@ -34,19 +34,19 @@ def pull_manifest(self: Task, *, mtn: List[str], username: str) -> dict:
 
 @shared_task(name="sign manifests", bind=True, acks_late=True)
 def sign_manifest(
-    self: Task,
-    *,
-    username: str,
-    **signature_data: dict,
+        self: Task,
+        *,
+        username: str,
+        **signature_data: dict,
 ) -> Dict:
     """
     a task to Quicker Sign manifest, by MTN, in RCRAInfo
     """
-    from apps.trak.services import ManifestService
+    from apps.trak.services import EManifest
 
     try:
-        manifest_service = ManifestService(username=username)
-        return manifest_service.quicker_sign_manifests(**signature_data)
+        emanifest = EManifest(username=username)
+        return emanifest.submit_quick_signature(**signature_data)
     except (ConnectionError, TimeoutError) as exc:
         raise Reject(exc)  # To Do: add retry logic
     except Exception as exc:
@@ -77,18 +77,18 @@ def save_rcrainfo_manifest(self, *, manifest_data: dict, username: str):
     user who is creating the manifest
     """
     from apps.core.services import TaskService
-    from apps.trak.services import ManifestService, ManifestServiceError
+    from apps.trak.services import EManifest, EManifestError
 
     logger.info(f"start task: {self.name}")
     task_status = TaskService(task_id=self.request.id, task_name=self.name, status="STARTED")
     try:
-        manifest = ManifestService(username=username)
-        new_manifest = manifest.save_to_rcrainfo(manifest=manifest_data)
+        emanifest = EManifest(username=username)
+        new_manifest = emanifest.save(manifest=manifest_data)
         if new_manifest:
             task_status.update_task_status(status="SUCCESS", results=new_manifest)
             return new_manifest
-        raise ManifestServiceError("error creating manifest")
-    except ManifestServiceError as exc:
+        raise EManifestError("error creating manifest")
+    except EManifestError as exc:
         logger.error(f"failed to create manifest ({manifest_data}): {exc.message}")
         task_status.update_task_status(status="FAILURE", results=exc.message)
         return {"error": exc.message}
