+++ server/apps/trak/views/manifest_view.py
@@ -11,8 +11,8 @@ from apps.sites.services import HaztrakSiteService
 from apps.trak.models import Manifest
 from apps.trak.serializers import ManifestSerializer
 from apps.trak.serializers.signature_serializer import QuickerSignSerializer
-from apps.trak.services import ManifestService
-from apps.trak.services.manifest_services import TaskResponse, get_manifests
+from apps.trak.services import EManifest, TaskResponse
+from apps.trak.services.manifest_services import get_manifests
 
 logger = logging.getLogger(__name__)
 
@@ -44,8 +44,8 @@ class CreateManifestView(GenericAPIView):
     def post(self, request: Request) -> Response:
         manifest_serializer = self.serializer_class(data=request.data)
         manifest_serializer.is_valid(raise_exception=True)
-        manifest = ManifestService(username=str(request.user))
-        data = manifest.create_manifest(manifest=manifest_serializer.data)
+        emanifest = EManifest(username=str(request.user))
+        data = emanifest.create(manifest=manifest_serializer.data)
         return Response(data=data, status=status.HTTP_201_CREATED)
 
 
@@ -109,8 +109,8 @@ class SignManifestView(GenericAPIView):
         quicker_serializer = self.serializer_class(data=request.data)
         quicker_serializer.is_valid(raise_exception=True)
         signature = quicker_serializer.save()
-        manifest = ManifestService(username=str(request.user))
-        data: TaskResponse = manifest.sign_manifests(signature=signature)
+        emanifest = EManifest(username=str(request.user))
+        data: TaskResponse = emanifest.sign(signature=signature)
         return Response(data=data, status=status.HTTP_200_OK)
